requirements.txt
aiofiles==23.2.1
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
archspec @ file:///croot/archspec_1709217642129/work
blinker==1.8.2
boltons @ file:///croot/boltons_1677628692245/work
cachetools==5.5.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1729155152784/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
distro @ file:///croot/distro_1714488253808/work
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.3
fastapi-cli==0.0.5
fasttext @ file:///home/conda/feedstock_root/build_artifacts/fasttext_1716087782544/work
ffmpy==0.4.0
filelock==3.16.1
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.3.0
gradio_client==1.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
idna @ file:///croot/idna_1714398848350/work
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libmambapy @ file:///croot/mamba-split_1725563367192/work/libmambapy
logdna==1.18.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1723567589013/work
multidict==6.0.4
narwhals==1.10.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1728239947573/work/dist/numpy-2.1.2-cp310-cp310-linux_x86_64.whl#sha256=cec0873aee25a2daf091c356544671803f9c6db7184e6d74d10fea0d48b30941
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
orjson==3.10.10
packaging @ file:///croot/packaging_1720101850331/work
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
protobuf==5.28.3
pyarrow==17.0.0
pybind11 @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1729256490124/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1729256490124/work
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
referencing==0.35.1
requests @ file:///croot/requests_1721410876868/work
rich==13.9.3
rpds-py==0.20.0
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
ruff==0.7.0
semantic-version==2.10.0
sentry-sdk==2.17.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
streamlit==1.39.0
stripe==11.1.1
termcolor==2.3.0
tiktoken==0.4.0
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.1
tqdm @ file:///croot/tqdm_1724853939799/work
truststore @ file:///croot/truststore_1695244293384/work
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3 @ file:///croot/urllib3_1700840488771/work
uvicorn==0.32.0
uvloop==0.21.0
watchdog==5.0.3
watchfiles==0.24.0
websockets==12.0
zstandard @ file:///croot/zstandard_1728569189425/work
